chore: removing panics
